[maximal_planar] move Phase1 to own file
