+++ src/embeding/maximal_planar/mod.rs
@@ -1,1 +1,4 @@
 pub mod index;
+mod phase1;
+mod phase2;
+mod phase3;
